refactor: ActionMenu 공용 컴포넌트로 이동
