+++ src/components/Atoms/index.ts
@@ -1,3 +1,4 @@
+export { ActionMenu, type ActionMenuProps } from './ActionMenu/ActionMenu';
 export { Badge } from './Badge';
 export { Button } from './Button';
 export { Chip } from './Chip';
