+++ src/components/marketplace/index.ts
@@ -4,5 +4,5 @@ export { DetailContentFooter } from '@/components/marketplace/DetailContentFoote
 export { CommentSection } from '@/components/marketplace/comment/CommentSection';
 export { DetailAside } from '@/components/marketplace/DetailAside/DetailAside';
 export { ShoppingMeetingRegisterModal } from '@/components/marketplace/registerModal/MeetingRegisterModal';
-export { ActionMenu } from '@/components/marketplace/ActionMenu/ActionMenu';
+export { ActionMenu } from '@/components/Atoms/ActionMenu/ActionMenu';
 export { IntroSection } from '@/components/marketplace/IntroSection';
